api/requirements.txt
flask==2.3.3
google-generativeai==0.4.0
python-dotenv==1.0.0
Flask-Cors==4.0.0 